+++ src/main/scala/services/Question2.scala
@@ -8,7 +8,7 @@ object Question2 extends FinalResponse[Map[String,TransactionTypes]]{
 
   type TransactionByCategory = Map[String, List[Transaction]]
 
-  def getTransactionsByUserAndCategory(transactions: List[Transaction]): Map[String, Map[String, BigDecimal]] = {
+  def getTransactionsByUserAndCategory(transactions: List[Transaction]): Map[String, Map[String, Double]] = {
     transactions.groupBy(_.accountId).map {
       case (accountId, transaction) => accountId ->
         transaction.groupBy(_.category)
@@ -18,17 +18,17 @@ object Question2 extends FinalResponse[Map[String,TransactionTypes]]{
     }
   }
 
-  def calculateAverage(data: List[Transaction]): BigDecimal = {
+  def calculateAverage(data: List[Transaction]): Double = {
     val average = data.collect(_.transactionAmount).sum / data.length
-    BigDecimal.valueOf(average)
+    average
   }
 
-  def convertToTransactionTypesCaseClass(data: Map[String, Map[String, BigDecimal]]): Map[String, TransactionTypes] =
+  def convertToTransactionTypesCaseClass(data: Map[String, Map[String, Double]]): Map[String, TransactionTypes] =
     data.map {
       case (accountId, transaction) => accountId -> unApply(transaction)
     }
 
-  def unApply(ts: Map[String,BigDecimal]): TransactionTypes =
+  def unApply(ts: Map[String,Double]): TransactionTypes =
     TransactionTypes(ts.getOrElse("AA",0), ts.getOrElse("BB",0),ts.getOrElse("CC",0),ts.getOrElse("DD",0),ts.getOrElse("EE",0),ts.getOrElse("FF",0),ts.getOrElse("GG",0))
 
   override def exerciseSolver(transactions: List[Transaction]): Map[String, TransactionTypes] = convertToTransactionTypesCaseClass(getTransactionsByUserAndCategory(transactions))
